pyproject.toml
[tool.poetry]
name = "bookinghottelroom"
version = "0.1.0"
description = "Booking a Hottel Room"
authors = ["Dima Stantsel <dimastantsel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djoser = "^2.1.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
